fixed create file bug for azure pipeline

+++ redrum-not-muckduck-game/Game.cs
@@ -117,11 +117,15 @@ namespace redrum_not_muckduck_game
             SaveHints.GetWorkingHintDirectory();
 
             //Create files
-            File.Create(SaveVisitedRooms.WorkingVisitedRoomsDirectory);
-            File.Create(SaveHintQuotes.WorkingHintQuotesDirectory);
-            File.Create(SaveWholeBoard.WorkingBoardDirectory);
-            File.Create(SaveElements.WorkingElementDirectory);
-            File.Create(SaveHints.WorkingHintDirectory);
+            if (!File.Exists(SaveWholeBoard.WorkingBoardDirectory))
+            {
+                File.Create(SaveVisitedRooms.WorkingVisitedRoomsDirectory);
+                File.Create(SaveHintQuotes.WorkingHintQuotesDirectory);
+                File.Create(SaveWholeBoard.WorkingBoardDirectory);
+                File.Create(SaveElements.WorkingElementDirectory);
+                File.Create(SaveHints.WorkingHintDirectory);
+            }
+
 
             //If there is saved data - load it
             if (new FileInfo(SaveWholeBoard.WorkingBoardDirectory).Length != 0)
